[Deribit] Use primitive types and BigDecimal only

+++ xchange-deribit/src/main/java/org/knowm/xchange/deribit/v1/dto/marketdata/DeribitOrder.java
@@ -3,13 +3,15 @@ package org.knowm.xchange.deribit.v1.dto.marketdata;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import java.math.BigDecimal;
+
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class DeribitOrder {
 
     @JsonProperty("quantity") public int quantity;
     @JsonProperty("amount") public int amount;
-    @JsonProperty("price") public float price;
+    @JsonProperty("price") public BigDecimal price;
     @JsonProperty("cm") public int cm;
     @JsonProperty("cm_amount") public int cmAmount;
     @JsonProperty("oid") public String oid;
@@ -23,7 +25,7 @@ public class DeribitOrder {
         return amount;
     }
 
-    public float getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
